--- libraries/apollo-gradle-plugin/testProjects/operationIds/apollo-compiler-plugin/build.gradle.kts
@@ -1,9 +1,0 @@
-
-plugins {
-    alias(libs.plugins.kotlin.jvm)
-    alias(libs.plugins.apollo)
-}
-
-dependencies {
-    implementation("com.apollographql.apollo3:apollo-compiler")
-}
