--- libraries/apollo-gradle-plugin/testProjects/operationIds/apollo-compiler-plugin/src/main/resources/META-INF/services/com.apollographql.apollo3.compiler.Plugin
@@ -1,1 +1,0 @@
-apollo.plugin.MyPlugin
