--- libraries/apollo-gradle-plugin/testProjects/operationIds/app/build.gradle.kts
@@ -1,12 +1,0 @@
-plugins {
-  alias(libs.plugins.kotlin.jvm)
-  alias(libs.plugins.apollo)
-}
-
-apollo {
-  service("service") {
-
-    packageNamesFromFilePaths()
-    plugin(project(":apollo-compiler-plugin"))
-  }
-}
