+++ libraries/apollo-gradle-plugin/testProjects/operationIds/settings.gradle.kts
@@ -1,3 +1,1 @@
 apply(from = "../../../../gradle/test.settings.gradle.kts")
-include(":app")
-include(":apollo-compiler-plugin")
