+++ libraries/apollo-gradle-plugin/testProjects/operationIds/src/main/graphql/com/example/TestQuery.graphql
